Fixes for tests and refactor
