+++ features/checkout/shipping_order/seeing_shipping_method_when_all_units_match_to_shipping_category.feature
@@ -23,7 +23,7 @@ Feature: Seeing shipping methods which category is same as category of all my un
         And this shipping method requires that all units match to "Over-sized" shipping category
         And I am a logged in customer
 
-    @ui @api
+    @ui
     Scenario: Seeing only shipping method which category is same as categories of all my units
         Given I have product "Rocket T-shirt" in the cart
         And I have product "Picasso T-shirt" in the cart
@@ -44,7 +44,7 @@ Feature: Seeing shipping methods which category is same as category of all my un
         And I should see "Invisible Post" shipping method
         And I should not see "Raven Post" shipping method
 
-    @ui @api
+    @ui
     Scenario: Seeing no shipping methods if my units matches to different shipping categories
         And I have product "Rocket T-shirt" in the cart
         And I have product "Star Trek Ship" in the cart
