+++ features/checkout/shipping_order/seeing_shipping_method_when_no_units_match_to_shipping_category.feature
@@ -35,7 +35,7 @@ Feature: Seeing shipping methods which category is not same as any category of a
         And I complete the addressing step
         Then there should be information about no available shipping methods
 
-    @ui @api
+    @ui
     Scenario: Seeing no shipping methods if its category is same as one category from my units categories
         Given I have product "Picasso T-shirt" in the cart
         And I have product "Star Trek Ship" in the cart
@@ -44,7 +44,7 @@ Feature: Seeing shipping methods which category is not same as any category of a
         And I complete the addressing step
         Then there should be information about no available shipping methods
 
-    @ui @api
+    @ui
     Scenario: Seeing no shipping methods if any of my unit has product variants which shipping category matching to the shipping category from shipping method
         And the "T-shirt banana" product's "S" size belongs to "Over-sized" shipping category
         And I have "T-shirt banana" with Size "S" in the cart
@@ -54,7 +54,7 @@ Feature: Seeing shipping methods which category is not same as any category of a
         And I complete the addressing step
         Then there should be information about no available shipping methods
 
-    @ui @api
+    @ui
     Scenario: Seeing shipping methods if none of my unit has variant with shipping method matching to the shipping category from shipping method
         And I have "T-shirt banana" with Size "S" in the cart
         And I have "T-shirt banana" with Size "M" in the cart
