+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/CartShippingMethod.xml
@@ -2,7 +2,7 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="https://api-platform.com/schema/metadata https://api-platform.com/schema/metadata/metadata-2.5.xsd"
 >
-    <resource class="Sylius\Bundle\ApiBundle\DTO\CartShippingMethod" shortName="CartShippingMethod">
+    <resource class="Sylius\Bundle\ApiBundle\View\CartShippingMethod" shortName="CartShippingMethod">
         <attribute name="validation_groups">sylius</attribute>
 
         <property name="code" identifier="true" />
