+++ src/Sylius/Bundle/ApiBundle/Resources/config/serialization/View/CartShippingMethod.xml
@@ -15,7 +15,7 @@
             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
             xsi:schemaLocation="http://symfony.com/schema/dic/serializer-mapping https://symfony.com/schema/dic/serializer-mapping/serializer-mapping-1.0.xsd"
 >
-    <class name="Sylius\Bundle\ApiBundle\DTO\CartShippingMethod">
+    <class name="Sylius\Bundle\ApiBundle\View\CartShippingMethod">
         <attribute name="code">
             <group>shipping_method:read</group>
         </attribute>
