+++ src/Sylius/Bundle/ApiBundle/View/Factory/CartShippingMethodFactoryInterface.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\DTO\Factory;
+namespace Sylius\Bundle\ApiBundle\View\Factory;
 
-use Sylius\Bundle\ApiBundle\DTO\CartShippingMethodInterface;
+use Sylius\Bundle\ApiBundle\View\CartShippingMethodInterface;
 use Sylius\Component\Core\Model\ShippingMethodInterface;
 
 interface CartShippingMethodFactoryInterface
